Fix up the logging options to actually work. Omit stacks on warnings and informational messages.
git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@811 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/RunActivity/_Main.cs
@@ -30,8 +30,10 @@ using System.Linq;
 using System.Windows.Forms;
 using ORTS.Menu;
 
-namespace ORTS {
-    static class Program {
+namespace ORTS
+{
+    static class Program
+    {
         public static string[] Arguments;
         public static string Version;         // ie "0.6.1"
         public static string Build;           // ie "0.0.3661.19322 Sat 01/09/2010  10:44 AM"
@@ -39,7 +41,7 @@ namespace ORTS {
         public static string UserDataFolder;  // ie @"C:\Users\Wayne\AppData\Roaming\Open Rails"
         public static Random Random = new Random();  // primary random number generator used throughout the program
         public static Simulator Simulator;
-        private static Viewer3D Viewer;       // Must stay private to ensure Simulator and Viewer do not conflict - see http://www.elvastower.com/forums/index.php?/topic/12105-multiprocessing-in-open-rails/ 
+        private static Viewer3D Viewer;
 #if DEBUG_VIEWER
 		private static DebugViewerForm DebugViewer;
 #endif
@@ -47,7 +49,8 @@ namespace ORTS {
         /// <summary>
         /// The main entry point for the application.
         /// </summary>
-        static void Main(string[] args) {
+        static void Main(string[] args)
+        {
             InitBuildRevision();
 
             UserDataFolder = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), Application.ProductName);
@@ -76,7 +79,8 @@ namespace ORTS {
             InitLogging(settings);
 
             // Do the action specified or write out some help.
-            switch (action) {
+            switch (action)
+            {
                 case "start":
                 case "start-profile":
                     Start(settings, data);
@@ -100,8 +104,10 @@ namespace ORTS {
         /// <summary>
         /// Run the specified activity from the beginning.
         /// </summary>
-        static void Start(UserSettings settings, string[] args) {
-            Action start = () => {
+        static void Start(UserSettings settings, string[] args)
+        {
+            Action start = () =>
+            {
                 InitSimulator(settings, args);
                 Simulator.Start();
                 Viewer = new Viewer3D(Simulator);
@@ -121,12 +127,18 @@ namespace ORTS {
 				DebugViewer.Dispose();
 #endif
             };
-            if (Debugger.IsAttached) {
+            if (Debugger.IsAttached)
+            {
                 start();
-            } else {
-                try {
+            }
+            else
+            {
+                try
+                {
                     start();
-                } catch (Exception error) {
+                }
+                catch (Exception error)
+                {
                     Trace.WriteLine(error);
                     if (settings.ShowErrorDialogs)
                         MessageBox.Show(error.ToString(), Application.ProductName);
@@ -140,9 +152,12 @@ namespace ORTS {
         /// the users local program storage, 
         /// ie.  "C:\\Users\\Wayne\\AppData\\Roaming\\ORTS\\SAVE.BIN"
         /// </summary>
-        public static void Save() {
-            Action save = () => {
-                using (BinaryWriter outf = new BinaryWriter(new FileStream(UserDataFolder + "\\SAVE.BIN", FileMode.Create, FileAccess.Write))) {
+        public static void Save()
+        {
+            Action save = () =>
+            {
+                using (BinaryWriter outf = new BinaryWriter(new FileStream(UserDataFolder + "\\SAVE.BIN", FileMode.Create, FileAccess.Write)))
+                {
                     // Save some version identifiers so we can validate on load.
                     outf.Write(Version);
                     outf.Write(Build);
@@ -157,12 +172,18 @@ namespace ORTS {
                     Console.WriteLine();
                 }
             };
-            if (Debugger.IsAttached) {
+            if (Debugger.IsAttached)
+            {
                 save();
-            } else {
-                try {
+            }
+            else
+            {
+                try
+                {
                     save();
-                } catch (Exception error) {
+                }
+                catch (Exception error)
+                {
                     Trace.WriteLine(error);
                     if (Simulator.Settings.ShowErrorDialogs)
                         MessageBox.Show(error.ToString(), Application.ProductName);
@@ -173,20 +194,26 @@ namespace ORTS {
         /// <summary>
         /// Resume a saved game.
         /// </summary>
-        static void Resume(UserSettings settings, string[] args) {
-            Action resume = () => {
+        static void Resume(UserSettings settings, string[] args)
+        {
+            Action resume = () =>
+            {
                 var saveFile = args.Length == 0 ? UserDataFolder + "\\SAVE.BIN" : args[0];
-                using (BinaryReader inf = new BinaryReader(new FileStream(saveFile, FileMode.Open, FileAccess.Read))) {
+                using (BinaryReader inf = new BinaryReader(new FileStream(saveFile, FileMode.Open, FileAccess.Read)))
+                {
                     // Read in validation data.
                     var revision = "<unknown>";
                     var build = "<unknown>";
                     var versionOkay = false;
-                    try {
+                    try
+                    {
                         revision = inf.ReadString().Replace("\0", "");
                         build = inf.ReadString().Replace("\0", "");
                         versionOkay = (revision == Version) && (build == Build);
-                    } catch { }
-                    if (!versionOkay) {
+                    }
+                    catch { }
+                    if (!versionOkay)
+                    {
                         if (revision.Length + build.Length > 0)
                             throw new InvalidDataException(String.Format("{0} save file is not compatible with V{1} ({2}); it was probably created by V{3} ({4}). Save files must be created by the same version of {0}.", Application.ProductName, Version, Build, revision, build));
                         throw new InvalidDataException(String.Format("{0} save file is not compatible with V{1} ({2}). Save files must be created by the same version of {0}.", Application.ProductName, Version, Build));
@@ -205,12 +232,18 @@ namespace ORTS {
                 }
                 Viewer.Run();
             };
-            if (Debugger.IsAttached) {
+            if (Debugger.IsAttached)
+            {
                 resume();
-            } else {
-                try {
+            }
+            else
+            {
+                try
+                {
                     resume();
-                } catch (Exception error) {
+                }
+                catch (Exception error)
+                {
                     Trace.WriteLine(error);
                     if (settings.ShowErrorDialogs)
                         MessageBox.Show(error.ToString(), Application.ProductName);
@@ -221,10 +254,12 @@ namespace ORTS {
         /// <summary>
         /// Tests OR against every activity in every route in every folder.
         /// </summary>
-        static void TestAll(string[] args) {
+        static void TestAll(string[] args)
+        {
             var settings = GetSettings(new[] { "ShowErrorDialogs=no", "Profiling", "ProfilingFrameCount=0" });
             InitLogging(settings);
-            Action testAll = () => {
+            Action testAll = () =>
+            {
                 var activities = (args.Length == 0 ? Folder.GetFolders() : args.Select(a => new Folder(Path.GetFileName(a), a)))
                     .SelectMany(f => Route.GetRoutes(f))
                     .SelectMany(r => ORTS.Menu.Activity.GetActivities(r))
@@ -232,7 +267,8 @@ namespace ORTS {
                     .OrderBy(a => a.FileName, StringComparer.OrdinalIgnoreCase)
                     .ToList();
                 var results = new bool[activities.Count];
-                Action<int> run = (i) => {
+                Action<int> run = (i) =>
+                {
                     InitSimulator(settings, new[] { activities[i].FileName }, "");
                     Simulator.Start();
                     Viewer = new Viewer3D(Simulator);
@@ -241,13 +277,20 @@ namespace ORTS {
                     results[i] = true;
                     Simulator.Stop();
                 };
-                for (var i = 0; i < activities.Count; i++) {
-                    if (Debugger.IsAttached) {
+                for (var i = 0; i < activities.Count; i++)
+                {
+                    if (Debugger.IsAttached)
+                    {
                         run(i);
-                    } else {
-                        try {
+                    }
+                    else
+                    {
+                        try
+                        {
                             run(i);
-                        } catch (Exception error) {
+                        }
+                        catch (Exception error)
+                        {
                             Trace.WriteLine(error);
                         }
                     }
@@ -261,18 +304,25 @@ namespace ORTS {
                 }
 
                 Console.WriteLine();
-                for (var i = 0; i < activities.Count; i++) {
+                for (var i = 0; i < activities.Count; i++)
+                {
                     Console.WriteLine("{0,-4}  {1}", results[i] ? "PASS" : "fail", activities[i].FileName);
                 }
                 Console.WriteLine();
                 Console.WriteLine("Tested {0} activities; {1} passed, {2} failed.", results.Length, results.Count(r => r), results.Count(r => !r));
             };
-            if (Debugger.IsAttached) {
+            if (Debugger.IsAttached)
+            {
                 testAll();
-            } else {
-                try {
+            }
+            else
+            {
+                try
+                {
                     testAll();
-                } catch (Exception error) {
+                }
+                catch (Exception error)
+                {
                     Trace.WriteLine(error);
                     if (settings.ShowErrorDialogs)
                         MessageBox.Show(error.ToString(), Application.ProductName);
@@ -280,13 +330,17 @@ namespace ORTS {
             }
         }
 
-        static void InitBuildRevision() {
-            try {
-                using (StreamReader f = new StreamReader("Version.txt")) {
+        static void InitBuildRevision()
+        {
+            try
+            {
+                using (StreamReader f = new StreamReader("Version.txt"))
+                {
                     Version = f.ReadLine();
                 }
 
-                using (StreamReader f = new StreamReader("Revision.txt")) {
+                using (StreamReader f = new StreamReader("Revision.txt"))
+                {
                     var line = f.ReadLine();
                     var revision = line.Substring(11, line.IndexOf('$', 11) - 11).Trim();
                     if (revision != "000")
@@ -298,37 +352,47 @@ namespace ORTS {
                     Build = Build + " " + f.ReadLine();  // date
                     Build = Build + " " + f.ReadLine(); // time
                 }
-            } catch {
+            }
+            catch
+            {
                 Version = "";
                 Build = Application.ProductVersion;
             }
         }
 
-        static UserSettings GetSettings(IEnumerable<string> options) {
+        static UserSettings GetSettings(IEnumerable<string> options)
+        {
             return new UserSettings(RegistryKey, options);
         }
 
-        static void InitLogging(UserSettings settings) {
+        static void InitLogging(UserSettings settings)
+        {
             var logFileName = "";
-            if ((settings.LoggingPath.Length > 0) && Directory.Exists(settings.LoggingPath)) {
-                var fileName = settings.LoggingFilename;
-                try {
-                    fileName = String.Format(fileName, Application.ProductName, Version.Length > 0 ? Version : Build, Version, Build, DateTime.Now);
-                } catch { }
-                foreach (var ch in Path.GetInvalidFileNameChars())
-                    fileName = fileName.Replace(ch, '.');
+            if (settings.Logging)
+            {
+                if ((settings.LoggingPath.Length > 0) && Directory.Exists(settings.LoggingPath))
+                {
+                    var fileName = settings.LoggingFilename;
+                    try
+                    {
+                        fileName = String.Format(fileName, Application.ProductName, Version.Length > 0 ? Version : Build, Version, Build, DateTime.Now);
+                    }
+                    catch { }
+                    foreach (var ch in Path.GetInvalidFileNameChars())
+                        fileName = fileName.Replace(ch, '.');
 
-                logFileName = Path.Combine(settings.LoggingPath, fileName);
-                // Ensure we start with an empty file.
-                File.Delete(logFileName);
-                // Make Console.Out go to the log file AND the output stream.
-                Console.SetOut(new FileTeeLogger(logFileName, Console.Out));
-                // Make Console.Error go to the new Console.Out.
-                Console.SetError(Console.Out);
+                    logFileName = Path.Combine(settings.LoggingPath, fileName);
+                    // Ensure we start with an empty file.
+                    File.Delete(logFileName);
+                    // Make Console.Out go to the log file AND the output stream.
+                    Console.SetOut(new FileTeeLogger(logFileName, Console.Out));
+                    // Make Console.Error go to the new Console.Out.
+                    Console.SetError(Console.Out);
+                }
             }
 
             // Captures Trace.Trace* calls and others and formats.
-            var traceListener = new ORTraceListener(Console.Out);
+            var traceListener = new ORTraceListener(Console.Out, !settings.Logging);
             traceListener.TraceOutputOptions = TraceOptions.Callstack;
             // Trace.Listeners and Debug.Listeners are the same list.
             Trace.Listeners.Add(traceListener);
@@ -342,17 +406,27 @@ namespace ORTS {
             LogSeparator();
             settings.Log();
             LogSeparator();
+            if (!settings.Logging)
+            {
+                Console.WriteLine("Logging is disabled, only fatal errors will appear here.");
+                LogSeparator();
+            }
         }
 
-        static void InitSimulator(UserSettings settings, string[] args) {
+        static void InitSimulator(UserSettings settings, string[] args)
+        {
             InitSimulator(settings, args, "");
         }
 
-        static void InitSimulator(UserSettings settings, string[] args, string mode) {
+        static void InitSimulator(UserSettings settings, string[] args, string mode)
+        {
             Console.WriteLine(mode.Length > 0 ? "Mode       = {0} {1}" : "Mode       = {1}", mode, args.Length == 1 ? "Activity" : "Explore");
-            if (args.Length == 1) {
+            if (args.Length == 1)
+            {
                 Console.WriteLine("Activity   = {0}", args[0]);
-            } else {
+            }
+            else
+            {
                 Console.WriteLine("Path       = {0}", args[0]);
                 Console.WriteLine("Consist    = {0}", args[1]);
                 Console.WriteLine("Time       = {0}", args[2]);
@@ -369,7 +443,8 @@ namespace ORTS {
                 Simulator.SetExplore(args[0], args[1], args[2], args[3], args[4]);
         }
 
-        static void LogSeparator() {
+        static void LogSeparator()
+        {
             Console.WriteLine(new String('-', 80));
         }
     }
